requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
https://github.com/ezyatev/django-admin-menu/archive/master.zip
django-attachments==1.4
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-extensions==2.1.6
django-js-asset==1.2.2
django-undeletable==1.1.0
idna==2.8
libsass==0.17.0
mock==2.0.0
pbr==5.1.3
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
sh==1.12.14
six==1.12.0
urllib3==1.24.1
django-import-export==1.2.0